Added Towerify to the "Self-hosting Solutions" section
# Description

This PR adds the Towerify app to the "Self-hosting Solutions" section of the /awesome-selfhosted/awesome-selfhosted repository.

The goal of Towerify is to efficiently manage a fleet of YunoHost servers, making it easier for administrators to deploy, configure, and monitor these servers. Towerify is designed to be flexible and can be used with YunoHost servers that are already set up and running. It is licensed under Elastic Licence 2.0 and developed using PHP and Bash.

# Changes Made

Added Towerify to the "Self-hosting Solutions" section.

Updated README.md to include a brief description, license information, and programming languages used.